Move: follow 관련 기능은 user 패키지로, style 관련 기능은 style 패키지로 이동
