+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/style/api/StyleController.kt
@@ -2,6 +2,7 @@ package com.wafflestudio.toyproject.team4.core.style.api
 
 import com.wafflestudio.toyproject.team4.common.Authenticated
 import com.wafflestudio.toyproject.team4.common.UserContext
+import com.wafflestudio.toyproject.team4.core.style.api.request.PostStyleRequest
 import com.wafflestudio.toyproject.team4.core.style.api.response.StyleResponse
 import com.wafflestudio.toyproject.team4.core.style.service.StyleService
 import com.wafflestudio.toyproject.team4.core.user.service.AuthTokenService
@@ -38,6 +39,12 @@ class StyleController(
         val username = authToken?.let { authTokenService.getUsernameFromToken(it) }
         return styleService.getStyle(username, styleId)
     }
+
+    @GetMapping("/user/{userId}/styles")
+    fun getUserStyles(
+        @PathVariable(value = "userId") userId: Long
+    ) = ResponseEntity(styleService.getUserStyles(userId), HttpStatus.OK)
+
 
     @Authenticated
     @PostMapping("/style")
