+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/style/api/response/UserStylesResponse.kt
@@ -1,7 +1,7 @@
-package com.wafflestudio.toyproject.team4.core.user.api.response
+package com.wafflestudio.toyproject.team4.core.style.api.response
 
 import com.wafflestudio.toyproject.team4.core.style.domain.Style
 
-data class StylesResponse(
+data class UserStylesResponse(
     val styles: List<Style.Preview>
 )
