+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/style/service/StyleService.kt
@@ -4,9 +4,10 @@ import com.wafflestudio.toyproject.team4.common.CustomHttp400
 import com.wafflestudio.toyproject.team4.common.CustomHttp404
 import com.wafflestudio.toyproject.team4.core.item.database.ItemEntity
 import com.wafflestudio.toyproject.team4.core.item.database.ItemRepository
-import com.wafflestudio.toyproject.team4.core.style.api.PostStyleRequest
+import com.wafflestudio.toyproject.team4.core.style.api.request.PostStyleRequest
 import com.wafflestudio.toyproject.team4.core.style.api.response.StyleResponse
 import com.wafflestudio.toyproject.team4.core.style.api.response.StylesResponse
+import com.wafflestudio.toyproject.team4.core.style.api.response.UserStylesResponse
 import com.wafflestudio.toyproject.team4.core.style.database.StyleEntity
 import com.wafflestudio.toyproject.team4.core.style.database.StyleRepository
 import com.wafflestudio.toyproject.team4.core.style.database.StyleRepositoryCustomImpl
@@ -15,12 +16,13 @@ import com.wafflestudio.toyproject.team4.core.user.database.UserRepository
 import com.wafflestudio.toyproject.team4.core.user.service.UserService
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
+import javax.transaction.Transactional
 import kotlin.math.ceil
 
 interface StyleService {
     fun getStyles(index: Long, count: Long, sort: String?): StylesResponse
     fun getStyle(username: String?, styleId: Long): StyleResponse
+    fun getUserStyles(userId: Long): UserStylesResponse
     fun postStyle(username: String, postStyleRequest: PostStyleRequest)
 }
 
@@ -70,6 +72,15 @@ class StyleServiceImpl(
             isLike = isLike,
         )
     }
+
+    @Transactional
+    override fun getUserStyles(userId: Long): UserStylesResponse {
+        val user = userRepository.findByIdOrNullWithStylesOrderByRecentDesc(userId)
+            ?: throw CustomHttp404("존재하지 않는 사용자입니다.")
+
+        return UserStylesResponse(user.styles.map { Style.preview(it) })
+    }
+
 
     private fun findStyleItems(style: StyleEntity): List<ItemEntity> {
         val styleItemIds = style.styleItems.map { it.itemId }
